Added generics for Wurfl
This one is a bit tricky, because there are many generics in use by
Wurfl.

The model reported by Wurfl is in certain cases an operating system
with version number. So I though it would be undesirable to include the
version number, because it would require this list to be updated with
every new OS version. So I just check if it starts with certain OS
names that are known to be generics.

I also didn’t add a check for all results, because that would cause the
actual OS to be flagged as a generic too.